# specs/contracts/api.yaml
openapi: 3.0.3
info:
  title: Aftertalk Core API
  description: REST API for Aftertalk Core Engine - Session management, transcription retrieval, and minutes operations
  version: 1.0.0
  contact:
    name: Aftertalk Team
servers:
  - url: https://api.aftertalk.io/v1
    description: Production server
  - url: https://api-staging.aftertalk.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /transcriptions:
    get:
      summary: List transcriptions
      description: Retrieve transcription segments filtered by session
      operationId: listTranscriptions
      tags:
        - Transcriptions
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Transcription segments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionList'
  /transcriptions/{transcriptionId}:
    get:
      summary: Get transcription by ID
      description: Retrieve one transcription segment
      operationId: getTranscription
      tags:
        - Transcriptions
      parameters:
        - name: transcriptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transcription segment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionSegment'
        '404':
          description: Transcription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
...
    TranscriptionList:
      type: object
      required:
        - transcriptions
        - total
        - limit
        - offset
      properties:
        transcriptions:
          type: array
          items:
            $ref: '#/components/schemas/TranscriptionSegment'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    TranscriptionSegment:
      type: object
      required:
        - id
        - session_id
        - segment_index
        - role
        - start_ms
        - end_ms
        - text
        - provider
        - status
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        segment_index:
          type: integer
        role:
          type: string
        start_ms:
          type: integer
          description: Milliseconds since session start
        end_ms:
          type: integer
          description: Milliseconds since session start
        text:
          type: string
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        provider:
          type: string
        language:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, processing, ready, error]
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true